Added AXIS_PORT Environment variable to Dockerfile Init script will write a file to prevent init to run on subsequent container starts

+++ config/init.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+if [ -f /opt/holodeck/bin/init.done ]; then exit 1; fi
+
 if [ "$AXIS_PORT" != "" ]; then
 	/usr/bin/xsltproc --param port $AXIS_PORT /opt/holodeck/bin/setWSPort.xsl /opt/holodeck/conf/axis2.xml > /opt/holodeck/conf/axis2_out.xml
 	mv /opt/holodeck/conf/axis2_out.xml /opt/holodeck/conf/axis2.xml
@@ -19,5 +21,7 @@ if [ "$ROLE" != "" ]; then
 		fi
 	fi
 fi
+
+/bin/touch /opt/holodeck/bin/init.done
 
 sh /opt/holodeck/bin/startServer.sh
